Replace planet requires with local requires of main.rkt.
Otherwise this package's implementation would be using the planet
package -- and causing users of this package to download the planet
package -- which doesn't seem right.

Furthermore, my understanding is that it's preferable within a package
to use a relative require -- like `(require "main.rkt")` -- instead of
requiring the collection itself -- like either of `(require describe)`
or `(require (planet williams/describe/describe))`. So even if this were
still a planet package, I think this would be a good change to make.